requirements.txt
torch
torchvision
pillow
kornia
transformers
streamlit
huggingface
timm
modelscope